feat: move `Pull.flare` to storage
This commit is the initial work (+ migration) to move `Pull.flare` to storage.
This column has crashed the DB in only 1 occasion in the past.

Notice that Pull doesn't have `commit` (because multiple commits refer to the same pull)
the issue with that is that in `Pull.get_commitid` I set it to return `"multiple_commits"`.
This should not affect the usage of `ArchiveField`, but will make it so that the data is
written to `/v4/repos/REPO_HASH/commits/multiple_commits/json_data/...`

It's fine by me, but we might want to consider another solution.

Part of codecov/platform-team#70